#43 update licence in cli epilog
